refactor(app): extract shared route-workspaces helpers from both routes
session-route.tsx and settings-route.tsx each carried a private copy of the
route-workspace helper family (RouteWorkspace, mapDesktopWorkspace,
workspaceLabel, mergeRouteWorkspaces, toSessionGroups, getSessionStatus,
isActiveSessionStatus, describe*Error, export/download helpers,
folderNameFromPath) — and the copies had drifted:

- settings-route was MISSING the mergeRouteWorkspaces fix that stops stale
  local-server registrations from clobbering desktop remote workspaces'
  routing fields (a latent settings-route bug, now fixed by unification)
- settings-route used older getSessionStatus (no normalizeSessionStatus)
  and isActiveSessionStatus (no 'streaming')

One canonical copy now lives in shell/route-workspaces.ts (the newer
session-route behavior). settings' toSessionGroups call passes an empty
loading set, preserving its previous output exactly.

session-route: 3,328 -> 3,161 lines; settings-route: 2,595 -> 2,472.

Tested:
- pnpm typecheck clean; bun test tests/ 58 pass
- pnpm build clean
- live: vite serves both routes' new module graph (200)
- pnpm test:e2e — 23/23 assertions ok (real app + opencode binary)
